Update CodeEditor and implement SwiftPoet library
This update integrates the SwiftPoet library in the CodeEditor module and modifies several classes to support this integration. Specifically, the XML file extension is enabled for the CodeLang class and language parsing is updated for the file reader. New functionalities to generate Swift code are added to the IconWriter and Svg2SwiftUI classes.