Refactor code and handle parsing exceptions
A refactor that includes an exception handling mechanism has been implemented in the CodeLang and Main files. The XML method in CodeLang has been simplified, and the allowed languages in the dropdown menu now load from an array in the Main file. Furthermore, an additional layer of safety has been added to code parser via a try-catch providing default behaviour in case any exception occurs.